src/requirements.txt
Flask==1.0.2
Jinja2==2.10
Markdown==2.6.11
pymemcache==2.1.0
PyYAML>=4.2b1
requests==2.20.0
# paramiko==2.4.1
suds-jurko==0.6
Flask-Cors==3.0.6
